The sub project wasm_hello_world_c_CMake_dll creates Native .so lib or WASM module.
See readme.md in the sub project folder. Not tested in Windows DLL yet.
I use 2 separate CMake projects here, one for the main app and another for the lib.
The main project is defined in the CMakeLists in the root folder.
The lib project is managed in the ```src/lib``` folder.